[Silver II] Title: 최소 힙, Time: 16 ms, Memory: 2916 KB -BaekjoonHub

+++ 백준/Silver/1927. 최소 힙/최소 힙.cc
@@ -5,10 +5,12 @@
 int main() {
 	std::ios::sync_with_stdio(0);
 	std::cin.tie(0);
-	std::priority_queue<long long, std::vector<long long>, std::greater<long long>> q;
+	//std::priority_queue<long long, std::vector<long long>, std::greater<long long>> q;
+	std::priority_queue<int, std::vector<int>, std::greater<int>> q;
+
 	int n;
 	std::cin >> n;
-	long long command;
+	int command;
 	for (int i = 0; i < n; i++) {
 		std::cin >> command;
 		if (command == 0) {
